Enhancement: New serializer module Removed serialization and deserialization methods from utlils/compression.h Added utils/serialization.h to handle serilization and deserialization of values
